[all] Changed all floating point types from Float to Double

+++ craftwerk/src/Craftwerk/Core/Color.hs
@@ -30,10 +30,10 @@ module Craftwerk.Core.Color (
   ) where
 
 -- | Color levels are in the range [0 .. 1]
-data Color = RGBA Float Float Float Float deriving (Show, Eq)
+data Color = RGBA Double Double Double Double deriving (Show, Eq)
 
 -- | Create a color from red, green and blue value. Color values are capped.
-makeColor :: Float -> Float -> Float -> Color
+makeColor :: Double -> Double -> Double -> Color
 makeColor r g b = RGBA (cap r) (cap g) (cap b) 1.0
 
 -- | Create a color from red, green and blue value in range [0 .. 100].
@@ -50,15 +50,15 @@ rgba r g b a = Just $ makeColorWithAlpha r g b a
 
 -- | Create a color from red, green, blue and alpha value.
 -- Color values are capped.
-makeColorWithAlpha :: Float -> Float -> Float -> Float -> Color
+makeColorWithAlpha :: Double -> Double -> Double -> Double -> Color
 makeColorWithAlpha r g b a = RGBA (cap r) (cap g) (cap b) (cap a)
 
 -- | Extract RGB values.
-getRGB :: Color -> (Float, Float, Float)
+getRGB :: Color -> (Double, Double, Double)
 getRGB (RGBA r g b a) = (r,g,b)
 
 -- | Extract RGBA values.
-getRGBA :: Color -> (Float, Float, Float, Float)
+getRGBA :: Color -> (Double, Double, Double, Double)
 getRGBA (RGBA r g b a) = (r,g,b,a)
 
 -- | Blend colors with integer percentage value (TODO: this is not as TikZ
@@ -68,9 +68,9 @@ getRGBA (RGBA r g b a) = (r,g,b,a)
 
 
 -- | Transform colors with a component wise map. Results get capped
-cmap :: (Float -> Float) -> Color -> Color
+cmap :: (Double -> Double) -> Color -> Color
 cmap f (RGBA r g b a) = RGBA (cap $ f r) (cap $ f g) (cap $ f b) a
 
-cap :: Float -> Float
+cap :: Double -> Double
 cap = min 1 . max 0
